common_includes.py
import sys
import os
import os.path
import json
import datetime

import aiofiles
from pathlib import Path
import glob

import psycopg2
import psycopg2.extras
from psycopg2 import sql

from fastapi import FastAPI, APIRouter, Request, Query, Depends, Body, File, UploadFile, status, Response
from fastapi.responses import FileResponse, JSONResponse
from fastapi.routing import APIRoute
from fastapi import BackgroundTasks
from typing import Optional, List, Dict, Callable

from starlette.types import Message
from starlette.requests import Request
from starlette.middleware.base import BaseHTTPMiddleware

from resources.models import file

from py_helpers.pytokens import *
from py_helpers.api_handler import *
from py_helpers.postgres_helper import *