requirements.txt
requests==2.26.0
fastapi==0.70.0
uvicorn==0.15.0
psutil==5.9.0
PyJWT==1.7.1
gunicorn==20.1.0
psycopg2-binary==2.9.2
python-multipart==0.0.5
icecream==2.1.3
aiofiles==0.8.0 
Jinja2==3.0.3
pytz==2019.3              
pytz-deprecation-shim==0.1.0.post0
